fix(DTFS2-7728): refactored test case
