+++ azure-functions/acbs-function/package.json
@@ -29,7 +29,7 @@
     "unit-test-ff": "jest --coverage --verbose --config=unit.ff.jest.config.js --passWithNoTests"
   },
   "dependencies": {
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "date-fns": "^2.30.0",
     "dotenv": "^16.4.5",
     "durable-functions": "^3.1.0"
