+++ dtfs-central-api/package.json
@@ -33,7 +33,7 @@
   },
   "dependencies": {
     "@ukef/dtfs2-common": "1.0.0",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "big.js": "^6.2.2",
     "compression": "^1.7.5",
     "date-fns": "^2.30.0",
@@ -55,7 +55,7 @@
     "tslib": "^2.8.1",
     "typeorm": "^0.3.20",
     "validator": "^13.12.0",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@types/compression": "^1.7.5",
@@ -68,11 +68,11 @@
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "jest": "29.5.0",
-    "jest-when": "^3.6.0",
-    "node-mocks-http": "^1.16.1",
+    "jest-when": "^3.7.0",
+    "node-mocks-http": "^1.16.2",
     "supertest": "6.3.3",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.2"
