+++ external-api/package.json
@@ -34,10 +34,10 @@
   "dependencies": {
     "@types/lodash": "^4.17.13",
     "@ukef/dtfs2-common": "1.0.0",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "compression": "^1.7.5",
     "cors": "^2.8.5",
-    "cron": "^3.2.1",
+    "cron": "^3.3.1",
     "date-fns": "^2.30.0",
     "dotenv": "16.3.1",
     "express": "^4.21.2",
@@ -72,11 +72,11 @@
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-airbnb-typescript-prettier": "^5.0.0",
     "eslint-config-prettier": "8.8.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "4.2.1",
     "jest": "29.5.0",
-    "node-mocks-http": "^1.16.1",
+    "node-mocks-http": "^1.16.2",
     "openapi-types": "^12.1.3",
     "prettier": "^2.8.8",
     "supertest": "^6.3.4",
