+++ gef-ui/package.json
@@ -37,7 +37,7 @@
     "@ministryofjustice/frontend": "3.0.2",
     "@ukef/dtfs2-common": "1.0.0",
     "ajv": "^8.17.1",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "chance": "^1.1.12",
     "cheerio": "1.0.0-rc.12",
     "clean-deep": "^3.4.0",
@@ -66,11 +66,11 @@
     "morgan": "1.10.0",
     "multer": "1.4.5-lts.1",
     "node-fetch": "^2.7.0",
-    "nodemon": "^3.1.7",
+    "nodemon": "^3.1.9",
     "nunjucks": "3.2.4",
     "path": "0.12.7",
     "redis": "^3.1.2",
-    "sanitize-html": "^2.13.1",
+    "sanitize-html": "^2.14.0",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
     "validator": "^13.12.0"
@@ -91,20 +91,20 @@
     "css-minimizer-webpack-plugin": "^5.0.1",
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "file-loader": "^6.2.0",
     "jest": "29.5.0",
     "jest-environment-jsdom": "^29.7.0",
-    "node-mocks-http": "^1.16.1",
+    "node-mocks-http": "^1.16.2",
     "rimraf": "^5.0.10",
-    "sass": "^1.81.0",
+    "sass": "^1.83.0",
     "sass-loader": "13.2.2",
     "style-loader": "3.3.2",
     "supertest": "^6.3.4",
-    "terser-webpack-plugin": "^5.3.10",
+    "terser-webpack-plugin": "^5.3.11",
     "ts-jest": "^29.2.5",
-    "webpack": "^5.96.1",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
     "webpack-merge": "^5.10.0"
   },
