+++ libs/common/package.json
@@ -47,7 +47,7 @@
     "unit-test-ff": "jest --coverage --verbose --config=unit.ff.jest.config.js --passWithNoTests"
   },
   "dependencies": {
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "big.js": "^6.2.2",
     "date-fns": "^3.3.1",
     "date-fns-tz": "^2.0.1",
@@ -58,10 +58,10 @@
     "mongodb": "^4.17.2",
     "mssql": "^10.0.4",
     "node-cron": "^3.0.3",
-    "node-mocks-http": "^1.16.1",
+    "node-mocks-http": "^1.16.2",
     "reflect-metadata": "^0.2.2",
     "typeorm": "^0.3.20",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
@@ -76,7 +76,7 @@
     "eslint": "^8.57.1",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
-    "jest-when": "^3.6.0",
+    "jest-when": "^3.7.0",
     "supertest": "^7.0.0",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.2"
