+++ libs/common/src/constants/tfm/index.ts
@@ -6,4 +6,4 @@ export * from './facility-stage';
 export * from './activity-types';
 export * from './comment-type';
 export * from './underwriter-manager-decisions';
-export * from './users';
+export * from './user-status';
