+++ package.json
@@ -64,7 +64,7 @@
     ]
   },
   "dependencies": {
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "express": "^4.21.2",
     "release-please": "^16.15.0"
   },
@@ -73,14 +73,14 @@
     "@commitlint/config-conventional": "^17.8.1",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
-    "cspell": "^8.16.1",
+    "cspell": "^8.17.1",
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "^5.2.1",
     "husky": "^9.1.7",
-    "lint-staged": "^15.2.10",
+    "lint-staged": "^15.2.11",
     "prettier": "3.0.3",
     "sort-package-json": "^2.12.0",
     "yaml-lint": "^1.7.0"
