+++ portal-api/package.json
@@ -34,7 +34,7 @@
   "dependencies": {
     "@azure/storage-file-share": "12.14.0",
     "@ukef/dtfs2-common": "1.0.0",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "comma-number": "^2.1.0",
     "compression": "^1.7.5",
     "cors": "2.8.5",
@@ -55,12 +55,12 @@
     "node-fetch": "^2.7.0",
     "passport": "0.6.0",
     "passport-jwt": "^4.0.1",
-    "sanitize-html": "^2.13.1",
+    "sanitize-html": "^2.14.0",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
     "validator": "^13.12.0",
     "xss": "^1.0.15",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
@@ -72,12 +72,12 @@
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "jest": "^29.7.0",
-    "jest-when": "^3.6.0",
+    "jest-when": "^3.7.0",
     "lodash": "^4.17.21",
-    "node-mocks-http": "^1.16.1",
+    "node-mocks-http": "^1.16.2",
     "prettier": "^2.8.8",
     "supertest": "6.3.3",
     "ts-jest": "^29.2.5",
