+++ portal/package.json
@@ -35,7 +35,7 @@
   "dependencies": {
     "@ministryofjustice/frontend": "3.0.2",
     "@ukef/dtfs2-common": "1.0.0",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "cheerio": "1.0.0-rc.12",
     "clamscan": "^2.4.0",
     "connect-flash": "0.1.1",
@@ -43,7 +43,7 @@
     "cookie-parser": "^1.4.7",
     "core-js": "^3.39.0",
     "csurf": "^1.11.0",
-    "csv-parser": "^3.0.0",
+    "csv-parser": "^3.1.0",
     "date-fns": "^2.30.0",
     "date-fns-tz": "^2.0.1",
     "dotenv": "16.0.3",
@@ -63,7 +63,7 @@
     "morgan": "1.10.0",
     "multer": "1.4.5-lts.1",
     "node-fetch": "^2.7.0",
-    "nodemon": "^3.1.7",
+    "nodemon": "^3.1.9",
     "nunjucks": "3.2.4",
     "path": "0.12.7",
     "redis": "^3.1.2",
@@ -92,20 +92,20 @@
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "file-loader": "^6.2.0",
     "jest": "29.5.0",
     "jest-environment-jsdom": "^29.7.0",
-    "jest-when": "^3.6.0",
-    "node-mocks-http": "^1.16.1",
+    "jest-when": "^3.7.0",
+    "node-mocks-http": "^1.16.2",
     "rimraf": "^5.0.10",
-    "sass": "^1.81.0",
+    "sass": "^1.83.0",
     "sass-loader": "13.2.2",
     "style-loader": "3.3.2",
-    "terser-webpack-plugin": "^5.3.10",
+    "terser-webpack-plugin": "^5.3.11",
     "ts-jest": "^29.2.5",
-    "webpack": "^5.96.1",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
     "webpack-merge": "^5.10.0"
   },
