+++ trade-finance-manager-api/api-tests/types/test-user.ts
@@ -3,7 +3,7 @@ import { TfmUser } from '@ukef/dtfs2-common';
 /**
  * Users returned by the test user fluent builder
  */
-export interface TestUser extends Pick<TfmUser, 'username' | 'teams' | 'email' | 'timezone' | 'lastName' | 'firstName'> {
+export interface TestUser extends Pick<TfmUser, 'username' | 'teams' | 'email' | 'timezone' | 'lastName' | 'firstName' | 'status'> {
   _id?: string;
   token: string;
   password: string;
