+++ trade-finance-manager-api/package.json
@@ -36,7 +36,7 @@
     "@babel/plugin-transform-runtime": "7.23.6",
     "@babel/preset-env": "7.23.6",
     "@ukef/dtfs2-common": "1.0.0",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "bson": "^4.7.2",
     "comma-number": "^2.1.0",
     "compression": "^1.7.5",
@@ -57,13 +57,13 @@
     "mongodb": "^4.17.2",
     "passport": "0.6.0",
     "passport-jwt": "^4.0.1",
-    "sanitize-html": "^2.13.1",
+    "sanitize-html": "^2.14.0",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^4.6.3",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
     "validator": "^13.12.0",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@types/compression": "^1.7.5",
@@ -76,11 +76,11 @@
     "axios-mock-adapter": "^2.1.0",
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "jest": "29.5.0",
-    "jest-when": "^3.6.0",
-    "node-mocks-http": "^1.16.1",
+    "jest-when": "^3.7.0",
+    "node-mocks-http": "^1.16.2",
     "supertest": "^6.3.4",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.2"
