--- trade-finance-manager-api/src/constants/user.js
@@ -1,8 +1,0 @@
-const STATUS = {
-  BLOCKED: 'blocked',
-  ACTIVE: 'active',
-};
-
-module.exports = {
-  STATUS,
-};
