+++ trade-finance-manager-api/src/v1/controllers/user/user.controller.js
@@ -2,10 +2,10 @@ const { ObjectId } = require('mongodb');
 const { generateAuditDatabaseRecordFromAuditDetails, deleteOne } = require('@ukef/dtfs2-common/change-stream');
 const { PAYLOAD_VERIFICATION, DocumentNotDeletedError } = require('@ukef/dtfs2-common');
 const { isVerifiedPayload } = require('@ukef/dtfs2-common/payload-verification');
+const { USER_STATUS } = require('@ukef/dtfs2-common');
 const { UserService } = require('../../services/user.service');
 const { mongoDbClient: db } = require('../../../drivers/db-client');
 const { mapUserData } = require('./helpers/mapUserData.helper');
-const { USER } = require('../../../constants');
 const utils = require('../../../utils/crypto.util');
 
 const businessRules = { loginFailureCount: 5 };
@@ -39,7 +39,7 @@ exports.create = async (user, auditDetails, callback) => {
   const collection = await db.getCollection('tfm-users');
   const tfmUser = {
     ...user,
-    status: USER.STATUS.ACTIVE,
+    status: USER_STATUS.ACTIVE,
     auditRecord: generateAuditDatabaseRecordFromAuditDetails(auditDetails),
   };
 
@@ -149,7 +149,7 @@ exports.incrementFailedLoginCount = async (user, auditDetails) => {
   const update = {
     loginFailureCount: failureCount,
     lastLoginFailure: Date.now(),
-    status: thresholdReached ? USER.STATUS.BLOCKED : user.status,
+    status: thresholdReached ? USER_STATUS.BLOCKED : user.status,
     auditRecord: generateAuditDatabaseRecordFromAuditDetails(auditDetails),
   };
