+++ trade-finance-manager-ui/package.json
@@ -39,7 +39,7 @@
     "@ministryofjustice/frontend": "3.0.2",
     "@ukef/dtfs2-common": "1.0.0",
     "ajv": "^8.17.1",
-    "axios": "^1.7.8",
+    "axios": "1.7.8",
     "chance": "^1.1.12",
     "cheerio": "1.0.0-rc.12",
     "chokidar": "^3.6.0",
@@ -70,13 +70,13 @@
     "nunjucks": "3.2.4",
     "path": "0.12.7",
     "redis": "^3.1.2",
-    "sanitize-html": "^2.13.1",
+    "sanitize-html": "^2.14.0",
     "supertest": "^6.3.4",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
     "validator": "^13.12.0",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@babel/core": "7.21.5",
@@ -102,19 +102,19 @@
     "css-minimizer-webpack-plugin": "^5.0.1",
     "eslint": "^8.57.1",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
     "file-loader": "^6.2.0",
     "jest": "29.5.0",
-    "jest-when": "^3.6.0",
-    "node-mocks-http": "^1.16.1",
-    "sass": "^1.81.0",
+    "jest-when": "^3.7.0",
+    "node-mocks-http": "^1.16.2",
+    "sass": "^1.83.0",
     "sass-loader": "13.2.2",
     "style-loader": "3.3.2",
-    "terser-webpack-plugin": "^5.3.10",
+    "terser-webpack-plugin": "^5.3.11",
     "ts-jest": "^29.2.5",
     "typescript": "^5.7.2",
-    "webpack": "^5.96.1",
+    "webpack": "^5.97.1",
     "webpack-cli": "^5.1.4",
     "webpack-merge": "^5.10.0"
   },
