+++ utils/package.json
@@ -22,9 +22,9 @@
     "dependencies": {
         "@faker-js/faker": "^8.4.1",
         "@ukef/dtfs2-common": "1.0.0",
-        "axios": "^1.7.8",
+        "axios": "1.7.8",
         "big.js": "^6.2.2",
-        "bson": "^6.10.0",
+        "bson": "^6.10.1",
         "chance": "^1.1.12",
         "date-fns": "^2.30.0",
         "dotenv": "^16.4.5",
@@ -39,7 +39,7 @@
         "reflect-metadata": "^0.2.2",
         "typeorm": "^0.3.20",
         "xml2js": "^0.5.0",
-        "zod": "^3.23.8"
+        "zod": "^3.24.1"
     },
     "devDependencies": {
         "@types/jsonwebtoken": "^9.0.7"
